retis-events: introduce python library.
The features used for pyo3 are different (and not-compatible)
for the embedded interpreter and python bindings, so move the
interpreter into an independent file that is easy to conditionally
compile off.

Use maturin to build with stable abi3.
