+++ retis-events/src/python.rs
@@ -3,13 +3,12 @@
 //! This module contains python bindings for retis events so that they can
 //! be inspected in post-processing tools written in python.
 
-use std::{collections::HashMap, fs, path::PathBuf, str::FromStr};
+use std::{collections::HashMap, path::PathBuf, str::FromStr};
 
-use anyhow::Result;
 use pyo3::{
     exceptions::{PyKeyError, PyRuntimeError},
     prelude::*,
-    types::{PyDict, PyList},
+    types::PyList,
 };
 
 use super::*;
@@ -208,7 +207,7 @@ impl PyEventSeries {
 /// ```
 #[pyclass(name = "EventReader")]
 pub(crate) struct PyEventReader {
-    factory: file::FileEventsFactory,
+    pub(crate) factory: file::FileEventsFactory,
 }
 
 #[pymethods]
@@ -383,46 +382,5 @@ pub(crate) fn to_pyobject(val: &serde_json::Value, py: Python<'_>) -> PyObject {
             let map: HashMap<_, _> = o.iter().map(|(k, v)| (k, to_pyobject(v, py))).collect();
             map.to_object(py)
         }
-    }
-}
-
-/// Create a python shell and execute the provided script.
-pub fn shell_execute(file: PathBuf, script: Option<&PathBuf>) -> Result<()> {
-    let event_file = PyEventFile::new(file)?;
-
-    Python::with_gil(|py| -> PyResult<()> {
-        let shell = PyShell::new(py, event_file)?;
-        if let Some(script) = script {
-            shell.run(&fs::read_to_string(script)?)
-        } else {
-            shell.interact()
-        }
-    })?;
-    Ok(())
-}
-
-/// Python shell.
-struct PyShell<'a> {
-    py: Python<'a>,
-    globals: Bound<'a, PyDict>,
-}
-
-impl<'a> PyShell<'a> {
-    const INTERACTIVE_SHELL: &'static str = "import code; code.interact(local=locals())";
-
-    fn new(py: Python<'a>, file: PyEventFile) -> PyResult<Self> {
-        let globals = PyDict::new_bound(py);
-        globals.set_item("reader", Py::new(py, file)?.into_bound(py))?;
-
-        Ok(Self { py, globals })
-    }
-
-    fn run(&self, script: &str) -> PyResult<()> {
-        self.py
-            .run_bound(script, Some(&self.globals.as_borrowed()), None)
-    }
-
-    fn interact(&self) -> PyResult<()> {
-        self.run(Self::INTERACTIVE_SHELL)
     }
 }
